refactor(core): keep lifecycle decision helpers internal
Move extracted determineStatus decision helpers into an internal module so the refactor does not expand the public lifecycle-manager API surface.

Also share merged and closed PR terminal-state mapping between cached and live SCM decision paths to prevent drift while addressing PR review feedback.
